refactor names, position button, structure project
