--- src/components/Greet.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-// function Greet() {
-//     return <p>Hello TK</p>
-// }
-
-const Greet = () => <p>hello tk</p>;
-
-export default Greet;
