+++ src/styles/animations.scss
@@ -1,10 +1,9 @@
 .App {
 	@include respond-below(xs) {
 		@media (prefers-reduced-motion: no-preference) {
-			animation: appScaleRv 500ms alternate 1 ease forwards;
-			// 	// animation-delay: 1s;
+			animation: appScale 500ms alternate 1 ease forwards;
 		}
-		@keyframes appScaleRv {
+		@keyframes appScale {
 			from {
 				transform: scale(1.09);
 			}
@@ -58,8 +57,8 @@
 	}
 	@include respond-above(md) {
 		#profile {
-			max-width: 768px;
-			margin: 0 auto;
+			// max-width: 768px;
+			// margin: 0 auto;
 		}
 	}
 }
