+++ src/styles/base.scss
@@ -19,7 +19,7 @@ body {
 	text-rendering: optimizeLegibility;
 	transition: background 0.2s linear;
 	a {
-		color: $color-blue;
+		color: $black;
 		text-decoration: none;
 	}
 }
@@ -29,7 +29,6 @@ body {
 	color: $color-light;
 
 	#logo {
-
 		a {
 			color: $green;
 		}
@@ -38,10 +37,9 @@ body {
 	a {
 		color: $color-green;
 	}
+
 	.profile__card {
-
 		.card__img {
-
 			#headshot {
 				border: 3px solid $grey3;
 			}
@@ -53,6 +51,11 @@ body {
 	ul {
 		list-style-type: none;
 	}
+
+	@include respond-above(xs) {
+		// margin: 0 auto;
+		// max-width: 768px;
+	}
 }
 
 header {
@@ -63,5 +66,6 @@ header {
 }
 
 // .wrapper_main {
-// 	margin: 0 1rem;
+// 	margin: 0 auto;
+// 	max-width: 768px;
 // }
